plugin: Split out the `struct channel_hint` handling
We're getting serious about how we manage the channel_hints, so let's
give them a proper home.
